+++ pyjoulescope_driver/entry_points/__init__.py
@@ -12,10 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from . import api_timeout, gpi, info, measure, mem_test, program, \
+from . import api_timeout, gpi, info, measure, mem_test, metadata, program, \
     record, scan, set_parameter, statistics, threads
 
-__all__ = [api_timeout, gpi, info, measure, mem_test, program,
+__all__ = [api_timeout, gpi, info, measure, mem_test, metadata, program,
            record, scan, set_parameter, statistics, threads]
 """This list of available command modules.  Each module must contain a 
 parser_config(subparser) function.  The function must return the callable(args)